server/requirements.txt
# NOTE: This file is for local development.
# Dockerfile uses split requirements files (requirements-*.txt) for better layer caching.
# If you update a dependency, update the corresponding split file too.

# Core ML / Math
numpy==1.26.4
scipy==1.10.1

# PyTorch (installed separately)
torch==2.0.1
torchaudio==2.0.2

# Audio
sounddevice==0.4.6
soundfile==0.12.1

# Speech
speechbrain==1.0.3
openai-whisper==20250625
pvporcupine==4.0.1

# HuggingFace (SpeechBrain deps)
huggingface_hub==0.16.4
hyperpyyaml==1.2.3
sentencepiece==0.1.99

# Server
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiofiles==23.2.1
requests==2.32.5
httpx==0.27.0

# LLM
groq>=0.11.0
llama-index>=0.10.0,<0.12.0
droidrun
